fix(ios): split announcement API to satisfy SwiftLint file length
